Cli cleanup (#81)
* cli cleanup and docs
move extractor and recompressor tools to separate modules, make more generic (not just for cli)
make cli tool setup more consistent
add readme description for extract tool
